refactor(sync-core): move discovery/ and providers.ts from server-elysia
These modules are framework-agnostic (pure TS, no Elysia deps) and belong
in sync-core so they can be consumed by any framework adapter (Elysia,
Hono, Cloudflare Workers).

- Moved discovery/awareness.ts and discovery/index.ts to sync-core
- Moved providers.ts to sync-core
- Added ./discovery subpath export to sync-core
- server-elysia re-exports from sync-core for backward compat
- Completed spec item 1.1
